Migration and seed: All left tables and seeders

+++ migrations/20220630113543_create_users_table.js
@@ -1,18 +1,22 @@
+const TABLE_NAME = 'users';
 /**
  * @param { import("knex").Knex } knex
  * @returns { Promise<void> }
  */
-exports.up = function(knex) {
-    return knex.schema.createTable('users', function(table) {
-        table.increments();
-        table.string('name').notNullable();
-        table.string('email').notNullable();
-        table.date('birthday').notNullable();
-        table.string('password').notNullable();
-        table.tinyint('gender').notNullable().defaultTo(0).references("value").inTable("gender").onDelete("CASCADE");
-        table.timestamp('created_at').defaultTo(knex.fn.now())
-        table.timestamp('updated_at').defaultTo(knex.fn.now())
-    });
+exports.up = async function(knex) {
+    if( !(await knex.schema.hasTable(TABLE_NAME)) ) {
+        return knex.schema.createTable(TABLE_NAME, (table) => {
+            table.increments();
+            table.string('name').notNullable();
+            table.string('email').notNullable();
+            table.date('birthday').notNullable();
+            table.string('password').notNullable();
+            table.tinyint('gender').notNullable().defaultTo(0).references("value").inTable("gender").onDelete("CASCADE");
+            table.timestamp('created_at').defaultTo(knex.fn.now())
+            table.timestamp('updated_at').defaultTo(knex.fn.now())
+        });
+    }
+    
 };
 
 /**
